Revamp of the data base.
Cards and CardEntities are mostly immutable. You have to change them through cardChanged().

The view model keeps the database and the card grouping. Empty slots are allowed in the card grouping and the game needs to implement isValid() to make check that if it is illegal.

The game is responsible for calculating the positions and groups of cards and handling user interactions. The viewmodel updates the in memory structures and the database.