Revamped Database
I changed the data base, mostly so the game state is not undoable.

The card table is basically the same. Cards and generatons where they changed are kept in the card table.

The state table is indexed by the game class name, except for common state which uses the view model class name. The state is blob that is marshalled from a Bundle.

There is a new highlight table that just keeps the cards that are currently highlighted. It was added to guarantee that we could restore a game if Android kicks the app out of memory.

Some methods were added to the CardDao to find the minimum generation and maximum generation and to trim generatoins that are not needed, if undo is tossed. The undo and redo buttons us this info to enable and disable themselves.

The view model is responsible for managing all of the database updates and keeps the card deck in card value order and the state lists used by the UI.

The launch icon was changed to a scorpion.