Simplified the Card Table
I realized that I could simplify things by removing the MutableState<Int> from Card and  folding CardEntity and Card into one class named Card, This works because when we change a card we update the card in the group list, which triggers a recompose. Once this simplification was done, Card.toEntity and Card.from, just became Card.copy. Note that the primary key which is only used to remove undo, is not declared in the Card constructor. Thus it is ignored in copy() and equals() and hashCode().