Revamp updating cardDeck and cardGroups
I changed this to update the cardDeck list and cardGroups in place from the changedCards and the highlightCards maps. Validity is checked throughout son the game can know the cards are OK.

If we find that the cardDeck or cardGroups is invalid we throw a DatabaseInconsistency exception. This is being handled in several places and is also used to abort database updates.

All updates to the cardDeck and cardGroups is through updateCards.

I added the card value to Card.toString()