Track all card and group positions
Card groups and card drawables now hold their offets and sizes. The hierarchy of rows and columns has been removed. The playarea size is maintained in the MainActivityViewModel by the MainActivity and is used by the games rather than the maxWidth and maxHeight of MainActivity BoxWithConstraints.

The game calculates the new offsets and sizes as cards are played. There are two functions to handle this, setupGroups() amd cardsIpdated(). SetupGroups sets the offsets of the card groups based on the playable area dimensions. It also also setsup other game specific values that may be needed. CardsUpdated() recalculates the card offsets withing each group and sets the sizes of each group. These methods are called when the cards are updated.

I ran into a bug where ths highlight wasn't be set properly. I belive it happened because a recomposition was overlapping with the highlight updates. I removed suspend from several methods to keep it from happening. A couple of methods wanted to update the game state in the database. To stop that, I look for game state changes in withUndo() and update the database there. The games just need to call onBundleUpdated() for this case. I also moved the highlight card update before the database update.

Fixed a bug where a king alone in a column could prevent detection of no more moves.