Drag cards on top of all other cards
I added a pass to CardGroup and CardDrawable. Cards are only drawn when the card group pass is the same as the card drawable pass. Cards that we are dragging are put in new groups with the drag pass without removing them from the main card groups. The drag card groups are draw and the end of the game content which puts them over everything.