Added autoscroll while dragging
To get this to work, I had to drag in the play area composable, or autoscroll would move the cards being dragged. So I added some coordinate translation to the DropCard interface, plus an onDrag method where the game can detect when autoscroll is needed and start it.

I refactored the scrolling code into a separate Scroller class that also handles autoscroll.