Removed double tap from Pyramid game
I added a Deal button to deal cards from the waste to stock groups, and added code to move cards from the stock to waste with a fling.

I also changed the deal count so the first one isn't counted.

I added a variant where the card on the stock pile is immediately moved to the waste when it is turned face up.

I changed the drag logic to use draggable2D so I can get the ending velocity to handle the fling.

I fixed or ignored several warnings.